+++ revdep/check.R
@@ -3,3 +3,5 @@ library("devtools")
 revdep_check(dependencies = c("Imports", "Depends"))
 revdep_check_save_summary()
 revdep_check_print_problems()
+
+revdep_email(date = "Mar 1")